+++ tests/settings.py
@@ -2,5 +2,5 @@ from os.path import join, dirname, abspath
 
 SAMPLES_DIR = join(dirname(dirname(abspath(__file__))), 'samples')
 SAMPLES_LIST_DIR = join(SAMPLES_DIR, 'list')
-SAMPLES_CONTENT_DIR = join(SAMPLES_DIR, 'content')
+SAMPLES_DETAIL_DIR = join(SAMPLES_DIR, 'detail')